+++ sqflite/test/version_test.dart
@@ -2,16 +2,17 @@
 
 import 'dart:io';
 
-import 'package:sentry/src/version.dart';
-import 'package:test/test.dart';
+import 'package:sentry_sqflite/src/version.dart';
+import 'package:flutter_test/flutter_test.dart';
 import 'package:yaml/yaml.dart' as yaml;
 
 void main() {
-  group('sdkVersion', () {
-    test('matches that of pubspec.yaml', () {
+  test(
+    'sdkVersion matches that of pubspec.yaml',
+    () {
       final dynamic pubspec =
           yaml.loadYaml(File('pubspec.yaml').readAsStringSync());
       expect(sdkVersion, pubspec['version']);
-    });
-  });
+    },
+  );
 }